+++ dbentry/site/views/base.py
@@ -154,6 +154,22 @@ class Inline:
 
     The inline will be rendered in tabular form if attribute `tabular` is True.
 
+    If the inline model is an M2M table, the inline will include template
+    context items `changelist_url` and `changelist_fk_field`. Together with some
+    javascript (added by InlineForm), this will add a link to the changelist
+    of the related model.
+
+    For most inlines, `changelist_url` and `changelist_fk_field` will be guessed
+    correctly. If the inline model has more than three model fields, you will
+    have to set `changelist_fk_field` attribute to the name of the ForeignKey
+    field to the target model.
+    Set `changelist_fk_field` to None to not have the link added.
+
+    If a `changelist_fk_field` can be guessed from the inline model, or if it
+    is set explicitly, the changelist will be filtered to only include the
+    items selected in the inline (the changelist_fk_field must be rendered with
+    a MIZSelect widget for this to work).
+
     The other attributes (model, form, fields, exclude, widgets) are arguments
     for the inline formset factory.
     """
@@ -163,10 +179,15 @@ class Inline:
     fields = None
     exclude = None
     widgets = None
+
     verbose_name = ""
     verbose_name_plural = ""
     verbose_model = None
+
     tabular = True
+
+    changelist_url = ""
+    changelist_fk_field = ""
 
     def __init__(self, parent_model):
         self.parent_model = parent_model
@@ -191,6 +212,8 @@ class Inline:
             "model_name": self.model._meta.model_name,
             "add_text": f"{self.verbose_name} hinzufügen",
             "tabular": self.tabular,
+            "changelist_url": self.get_changelist_url(),
+            "changelist_fk_field": self.get_changelist_fk_field(),
         }
 
     def get_formset_class(self):
@@ -208,6 +231,38 @@ class Inline:
             widgets=self.widgets,
             extra=1,
         )
+
+    def get_changelist_url(self):
+        """Return the URL for the changelist button for this inline."""
+        if self.changelist_url:
+            return self.changelist_url
+        if changelist_fk_field := self.get_changelist_fk_field():
+            field = self.model._meta.get_field(changelist_fk_field)
+            try:
+                return reverse(urlname("changelist", field.related_model._meta))
+            except NoReverseMatch:
+                return ""
+        else:
+            return ""
+
+    def get_changelist_fk_field(self):
+        """
+        Return the name of the field that contains the ids to be used in the
+        URL of the changelist button.
+        """
+        if self.changelist_fk_field is None:
+            # Do not create a changelist button.
+            return ""
+        if self.changelist_fk_field:
+            return self.changelist_fk_field
+        fields = self.model._meta.get_fields()
+        if len(fields) == 3:
+            # Can guess the field, assuming that the model contains these
+            # fields: ID field, FK to parent, FK to target model
+            for field in fields:
+                if field.is_relation and field.related_model != self.parent_model:
+                    return field.name
+        return ""
 
 
 class BaseEditView(