+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 
-- date picker for Ausgabe Erscheinungsdatum field
+- added date picker for Ausgabe Erscheinungsdatum field
+- added button to inlines that sends the user to the changelist of the items selected in the inline forms 
 
 ### Changed